fix: user.resto repo, renam Id struct to PK

+++ src/database/seeders/review.go
@@ -15,7 +15,7 @@ func ReviewSeeder() {
 	for i := 0; i < 10; i++ {
 		for j := 0; j < 10; j++ {
 			reviews = append(reviews, &models.Review{
-				Rating:       helpers.RandomElement([]float64{0.5, 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5}),
+				Rating:       helpers.RandomElement([]float32{0.5, 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5}),
 				UserId:       uint(j + 1),
 				RestaurantId: uint(i + 1),
 				Comment:      gofakeit.Comment(),
